# documents/OpenAPIDocs.yaml
openapi: 3.0.0
info:
  title: TaskBounty API
  description: API documentation for the TaskBounty application.
  version: 1.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/JustinVijar/TaskBounty/1.0.0
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /stripe/checkout/{bountyPostId}:
    get:
      summary: Create a checkout session
      description: Creates a Stripe checkout session for a bounty post. Requires authentication and ownership.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bountyPostId
          required: true
          schema:
            type: string
          description: The ID of the bounty post
      responses:
        '200':
          description: Checkout session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
        '401':
          description: Unauthorized (not authenticated or not owner)
        '404':
          description: Bounty post not found
  /stripe/payment_success/bounty_post:
    get:
      summary: Confirm payment for a bounty post
      description: Confirms payment for a bounty post after checkout.
      parameters:
        - in: query
          name: bountyPostId
          required: true
          schema:
            type: string
          description: The ID of the bounty post
        - in: query
          name: session_id
          required: true
          schema:
            type: string
          description: The Stripe session ID
      responses:
        '200':
          description: Payment confirmed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request (e.g., invalid parameters)
  /stripe/approve_solution/payout:
    post:
      summary: Approve a solution and initiate payout
      description: Approves a solution and initiates a payout. Requires authentication and ownership.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                solutionId:
                  type: string
              required:
                - solutionId
      responses:
        '200':
          description: Payout initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized (not authenticated or not owner)
        '404':
          description: Solution not found
  /stripe/approve_solution/transfer:
    post:
      summary: Approve a solution and initiate transfer
      description: Approves a solution and initiates a transfer. Requires authentication and ownership.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                solutionId:
                  type: string
              required:
                - solutionId
      responses:
        '200':
          description: Transfer initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized (not authenticated or not owner)
        '404':
          description: Solution not found
  /stripe/onboarding:
    get:
      summary: Onboard user to Stripe
      description: Provides a URL for Stripe onboarding. Requires authentication.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Onboarding URL provided successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
        '401':
          description: Unauthorized (not authenticated)
  /stripe/create_account:
    post:
      summary: Create a Stripe account for the user
      description: Creates a Stripe account for the user. Requires authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required:
                - email
      responses:
        '200':
          description: Stripe account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized (not authenticated)
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT